First push before ammending: Large progress in the black room and red room areas. Added a high complexity of options. To be further ammended.
Second push before ammending:
Removed all the prints and replaced them with messages and used the print_message function.
Added two new TODO's.
Added the 'no' option for the black door and choosing not to follow the blood trail.

Third push before ammending:
Solved some bugs pertaining to the order the rooms are chosen in, the elimination of options from the list once they are selected, added printing text letter by letter. Added options in case the green door is visited before the black door or the reverse. Now, only to make the transition to the third level from the puzzle corridor and the overall lower floor. Might rename the lower floor.